+++ components/ColorBlock.tsx
@@ -13,12 +13,12 @@ export default function (props: iColorBlock) {
   const { color, focusable, ...p } = props;
   return (
     <div
-      class="color-block m-0.5"
+      class="color-block h-inherit m-0.5"
       tabIndex={focusable ? 0 : undefined}
       {...p}
     >
       <div
-        class={`aspect-square rounded ${
+        class={`h-inherit aspect-square rounded ${
           color ? `bg-${color}` : "bg-obsidiana"
         } ${color === "blanco" || color === "obsidiana" ? "border-1" : null}`}
         title={color ? color : ""}