+++ components/ObsidianaCard.tsx
@@ -1,10 +1,10 @@
 import { ComponentChildren } from "preact";
 
-interface iPage {
+interface iObsidianaCard {
   children?: ComponentChildren;
 }
 
-export default function (props: iPage) {
+export default function (props: iObsidianaCard) {
   return (
     <div class="bg-obsidiana p-8 rounded clr-border border-1">
       {props.children}